+++ packages/tracing/src/integrations/index.ts
@@ -3,8 +3,8 @@ export { Postgres } from './node/postgres';
 export { Mysql } from './node/mysql';
 export { Mongo } from './node/mongo';
 export { Prisma } from './node/prisma';
-export { GraphQL } from './graphql';
-export { Apollo } from './apollo';
+export { GraphQL } from './node/graphql';
+export { Apollo } from './node/apollo';
 
 // TODO(v7): Remove this export
 // Please see `src/index.ts` for more details.
